chore : Modal 폴더 구조 및 불필요한 코드삭제, d.ts 파일 정리
