+++ src/components/Modal/index.tsx
@@ -1,18 +1,46 @@
-import { ReactNode } from "react";
-import Portal from "../Portal";
-import * as S from "./styled";
+import React, { ReactNode } from "react";
+import styled from "styled-components";
 
-export interface ModalProps {
-  children: ReactNode;
+interface ModalProps {
   isOpen: boolean;
+  closeModal?: () => void;
+  children: ReactNode;
 }
 
-function Modal({ isOpen, children }: ModalProps) {
+const Modal: React.FC<ModalProps> = ({ isOpen, closeModal, children }) => {
+  if (!isOpen) return null;
+
   return (
-    <Portal>
-      <S.Container $isOpen={isOpen}>{children}</S.Container>
-    </Portal>
+    <ModalBg onClick={closeModal}>
+      <Content onClick={(e) => e.stopPropagation()}>{children}</Content>
+    </ModalBg>
   );
-}
+};
 
 export default Modal;
+
+const ModalBg = styled.div`
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.5);
+  padding: 0 16px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 999;
+  max-width: 375px;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+`;
+
+const Content = styled.div`
+  background-color: white;
+  padding: 16px 20px;
+  border-radius: 8px;
+  width: 100%;
+  position: relative;
+`;
