--- src/types/balance.ts
@@ -1,10 +1,0 @@
-export interface Questions {
-  id: number;
-  description: string;
-  optionA: string;
-  optionB: string;
-}
-
-export interface SelectQuestions {
-  selectedOption: string;
-}
