--- src/types/situation.ts
@@ -1,6 +1,0 @@
-export interface Situation {
-  id: string;
-  title: string;
-  color: string;
-  textColor: string;
-}
