--- src/types/topic.ts
@@ -1,13 +1,0 @@
-export interface Topic {
-  id: number;
-  content: string;
-  isRecommend: boolean;
-  tips: TopicTip[];
-  situationType: string;
-  situationName: string;
-}
-
-export interface TopicTip {
-  title: string;
-  content: string;
-}
